# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Mesh:
    properties:
      address:
        type: string
        x-go-name: Address
      description:
        type: string
        x-go-name: Description
      id:
        description: the id for this mesh
        format: int64
        minimum: 0
        type: integer
        x-go-name: ID
      name:
        type: string
        x-go-name: Name
    required:
    - id
    type: object
    x-go-package: github.com/KrzysztofSieczkiewicz/go-service-template/src/data
info:
  description: '# Documentation for Product API'
  title: of Product API
  version: 1.0.0
paths:
  /meshes:
    get:
      description: Returns a list of meshes from the data store
      operationId: listMeshes
      responses:
        "200":
          $ref: '#/responses/meshesResponse'
      tags:
      - meshes
    post:
      description: Returns a 200 message
      operationId: postMeshes
      responses:
        "200":
          $ref: '#/responses/noContent'
      tags:
      - meshes
  /meshes/{id}:
    delete:
      description: Returns a 201 message
      operationId: deleteMesh
      parameters:
      - description: the id of the mesh to be deleted from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
      tags:
      - meshes
    put:
      description: Returns a 200 message
      operationId: putMesh
      parameters:
      - description: the id of the mesh to be deleted from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContent'
      tags:
      - meshes
produces:
- application/json
responses:
  meshesResponse:
    description: ""
    schema:
      items:
        $ref: '#/definitions/Mesh'
      type: array
  noContent:
    description: ""
schemes:
- http
swagger: "2.0"